Always store date in YYYY-MM-DD format
The format conversion is done at display.

WARNING: All the date type options will be reset on plugin upgrade

WARNING 2: The upgrade process will try to update the data to be in
YYYY-MM-DD. It may fail so be careful.

This patch also add a Format filter and a element_types_format function
designed to be used by other plugins to transform a raw value into a
format acceptable for the type plugin